add index.js files
