+++ src/components/BookList/index.js
@@ -1,0 +1,1 @@
+export { default as BookList } from './BookList';
