+++ src/components/Chapter/index.js
@@ -1,0 +1,1 @@
+export { default as Chapter } from './Chapter';
