+++ src/components/MenuBar/index.js
@@ -1,0 +1,1 @@
+export { default as MenuBar } from './MenuBar';
