+++ src/components/SubMenuBar/index.js
@@ -1,0 +1,1 @@
+export { default as SubMenuBar } from './SubMenuBar';
