+++ src/components/SupportTN/index.js
@@ -1,0 +1,1 @@
+export { default as SupportTN } from './SupportTN';
