+++ src/components/SupportTQ/index.js
@@ -1,0 +1,1 @@
+export { default as SupportTQ } from './SupportTQ';
