+++ src/components/index.js
@@ -1,0 +1,8 @@
+export * from './BookList';
+export * from './Card';
+export * from './Chapter';
+export * from './MenuBar';
+export * from './SubMenuBar';
+export * from './SupportTN';
+export * from './SupportTQ';
+export * from './TypoReport';
